Move include logic to dynamic_connectivity.hpp

+++ exercise3/implementation/dynamic_connectivity_sequential.hpp
@@ -8,7 +8,7 @@
 
 class DynamicConnectivity {
  public:
-    using Node = long;
+    using Node = int;
 
     DynamicConnectivity() = default;
     explicit DynamicConnectivity(long num_nodes);
